rework default logger
